Update player context mixin path for 1.18
The Palette stored in PalettedContainer is now stored inside a PalettedContainer$Data class instance, which is why access wideners are now needed to modify the variable. Additionally, information for the ChunkDataS2CPacket is now stored in the ChunkData class, so player context is passed in when the constructor is called.